+++ apps/web/src/app/customer/report/[id]/review/_components/ReviewFormSkeleton.tsx
@@ -1,13 +1,11 @@
-function Block({ className }: { className: string }) {
-  return <div className={`animate-pulse rounded-card bg-line-2 ${className}`} />;
-}
+import { SkeletonBlock } from "./SkeletonBlock";
 
 export function ReviewFormSkeleton() {
   return (
     <div className="flex flex-col gap-5">
-      <Block className="h-[4.5rem]" />
-      <Block className="h-[10rem]" />
-      <Block className="h-[14rem]" />
+      <SkeletonBlock className="h-[4.5rem]" />
+      <SkeletonBlock className="h-[10rem]" />
+      <SkeletonBlock className="h-[14rem]" />
     </div>
   );
 }
