+++ apps/web/src/app/customer/report/[id]/review/_components/SkeletonBlock.tsx
@@ -1,0 +1,3 @@
+export function SkeletonBlock({ className }: { className: string }) {
+  return <div className={`animate-pulse rounded-card bg-line-2 ${className}`} />;
+}
